fix: Move fonts file
