+++ src/redesign/style/index.jsx
@@ -1,6 +1,6 @@
 import colors from "./colors";
 import spacing from "../../style/spacing";
-import fonts from "./fonts";
+import fonts from "../../style/fonts";
 
 const style = {
   colors,
